Move DocRoute to own file
